Prevent background test poller cancellation
